+++ Application/RDD.Application/Controllers/AppController.cs
@@ -2,7 +2,9 @@
 using Rdd.Domain.Models.Querying;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
+using Rdd.Domain.Helpers;
 
 namespace Rdd.Application.Controllers
 {
@@ -21,62 +23,74 @@ namespace Rdd.Application.Controllers
         where TEntity : class, IEntityBase<TKey>
         where TKey : IEquatable<TKey>
     {
-        protected IUnitOfWork UnitOfWork { get; }
+        private readonly IUnitOfWork _unitOfWork;
 
         public AppController(IUnitOfWork unitOfWork, TCollection collection)
             : base(collection)
         {
-            UnitOfWork = unitOfWork;
+            _unitOfWork = unitOfWork;
         }
 
         public virtual async Task<TEntity> CreateAsync(ICandidate<TEntity, TKey> candidate, Query<TEntity> query)
         {
             var entity = await Collection.CreateAsync(candidate, query);
-
-            await UnitOfWork.SaveChangesAsync();
-
+            await SaveChangesAsync(entity.Yield());
             return entity;
         }
 
         public virtual async Task<IEnumerable<TEntity>> CreateAsync(IEnumerable<ICandidate<TEntity, TKey>> candidates, Query<TEntity> query)
         {
             var entities = await Collection.CreateAsync(candidates, query);
-
-            await UnitOfWork.SaveChangesAsync();
-
+            await SaveChangesAsync(entities);
             return entities;
         }
 
         public virtual async Task<TEntity> UpdateByIdAsync(TKey id, ICandidate<TEntity, TKey> candidate, Query<TEntity> query)
         {
             var entity = await Collection.UpdateByIdAsync(id, candidate, query);
-
-            await UnitOfWork.SaveChangesAsync();
-
+            await SaveChangesAsync(entity.Yield());
             return entity;
         }
 
         public virtual async Task<IEnumerable<TEntity>> UpdateByIdsAsync(IDictionary<TKey, ICandidate<TEntity, TKey>> candidatesByIds, Query<TEntity> query)
         {
             var entities = await Collection.UpdateByIdsAsync(candidatesByIds, query);
-
-            await UnitOfWork.SaveChangesAsync();
-
+            await SaveChangesAsync(entities);
             return entities;
         }
 
         public async Task DeleteByIdAsync(TKey id)
         {
             await Collection.DeleteByIdAsync(id);
-
-            await UnitOfWork.SaveChangesAsync();
+            await SaveChangesAsync(Enumerable.Empty<TEntity>());
         }
 
         public async Task DeleteByIdsAsync(IEnumerable<TKey> ids)
         {
             await Collection.DeleteByIdsAsync(ids);
-
-            await UnitOfWork.SaveChangesAsync();
+            await SaveChangesAsync(Enumerable.Empty<TEntity>());
         }
+
+        /// <summary>
+        /// Calls UnitOfWork.SaveChangesAsync() and pass modified items to OnBefore / OnAfter methods
+        /// </summary>
+        /// <param name="entities"></param>
+        /// <returns></returns>
+        protected virtual async Task SaveChangesAsync(IEnumerable<TEntity> entities)
+        {
+            await OnBeforeSaveEntitiesAsync(entities);
+            await _unitOfWork.SaveChangesAsync();
+            await OnAfterSaveEntitiesAsync(entities);
+        }
+
+        /// <summary>
+        /// Called before SaveChangesAsync(), last opportunity to modify entities
+        /// </summary>
+        protected virtual Task OnBeforeSaveEntitiesAsync(IEnumerable<TEntity> entities) => Task.CompletedTask;
+
+        /// <summary>
+        /// Called after SaveChangesAsync(), should be used to apply custom modifications before items are returned via API
+        /// </summary>
+        protected virtual Task OnAfterSaveEntitiesAsync(IEnumerable<TEntity> entities) => Task.CompletedTask;
     }
 }